separate plugin.xml and build.properties tests
