+++ plugins/com.google.cloud.tools.eclipse.test.util/src/com/google/cloud/tools/eclipse/test/util/BasePluginXmlTest.java
@@ -16,11 +16,6 @@
 
 package com.google.cloud.tools.eclipse.test.util;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.Set;
-
 import org.eclipse.core.runtime.IExtensionPoint;
 import org.eclipse.core.runtime.IExtensionRegistry;
 import org.eclipse.core.runtime.RegistryFactory;
@@ -35,8 +30,6 @@ import org.w3c.dom.NamedNodeMap;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 
-import com.google.common.collect.Sets;
-
 /**
  * Generic tests that should be true of all plugins.
  */
@@ -47,9 +40,7 @@ public abstract class BasePluginXmlTest {
   public final PluginXmlDocument pluginXmlDocument = new PluginXmlDocument();
   @Rule
   public final EclipseProperties pluginProperties = new EclipseProperties("plugin.properties");
-  @Rule
-  public final EclipseProperties buildProperties = new EclipseProperties("build.properties");
-  
+   
   private Document doc;
 
   @Before
@@ -94,34 +85,6 @@ public abstract class BasePluginXmlTest {
         Assert.assertNotNull("Null keyword " + id, keyword);
         Assert.assertFalse("Empty keyword " + id, keyword.isEmpty());
       }
-    }
-  }
-  
-  @Test
-  public final void testBuildProperties() throws IOException {
-    String[] binIncludes = buildProperties.get("bin.includes").split(",\\s*");
-    Set<String> includes = Sets.newHashSet(binIncludes);
-    
-    Assert.assertTrue(includes.contains("plugin.xml"));
-    Assert.assertTrue(includes.contains("plugin.properties"));
-    Assert.assertTrue(includes.contains("."));
-    Assert.assertTrue(includes.contains("META-INF/"));
-    
-    testIncludedIfPresent(includes, "helpContexts.xml");
-    testIncludedIfPresent(includes, "icons/");
-    testIncludedIfPresent(includes, "lib/");
-    testIncludedIfPresent(includes, "README.md");
-    testIncludedIfPresent(includes, "epl-v10.html");
-    testIncludedIfPresent(includes, "OSGI-INF/");
-    testIncludedIfPresent(includes, "fragment.xml");
-    testIncludedIfPresent(includes, "fragment.properties");
-  }
-
-  private static void testIncludedIfPresent(Set<String> includes, String name) 
-      throws IOException {
-    String path = EclipseProperties.getHostBundlePath() + "/" + name;
-    if (Files.exists(Paths.get(path))) {
-      Assert.assertTrue(includes.contains(name));
     }
   }
